CentrixSupport/Mental Health chatbot/Mental Health AI Chatbot/requirements.txt
opencv-python
python-docx
nltk
pdfplumber
playsound
pydub
pytesseract
openai-whisper
langchain
langchain-community
langchain-huggingface
langchain-ollama